Start building a VFS class to handle the VFS layer
